pyproject.toml
[project]
name = "ocr"
version = "0.1.0"
description = "Runs OCR of a book pages"
authors = [{name = "Tesla2000", email = "fratajczak124@gmail.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pre-commit>=3.7.1",
    "python-dotenv>=1.0.1",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.12.0",
    "google-cloud-vision>=3.8.0",
    "mypy>=1.13.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ocr"]

[dependency-groups]
google-drive = [
    "google-api-python-client>=2.188.0",
]
llm-cleanup = []
llm-cleanup-anthropic = [
    "anthropic>=0.76.0",
]
llm-cleanup-openai = [
    "openai>=2.15.0",
]
sylable-splitting = [
    "pyphen>=0.17.2",
]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
